Renamed to httpproxies
