+++ src/main/resources/lessons/httpproxies/documentation/10burp.adoc
