+++ src/main/resources/lessons/httpproxies/documentation/3browsersetup.adoc
