+++ src/main/resources/lessons/httpproxies/documentation/5configurefilterandbreakpoints.adoc
