+++ src/main/resources/lessons/httpproxies/documentation/6assignment.adoc
