+++ src/main/resources/lessons/httpproxies/documentation/7resend.adoc
