+++ src/main/resources/lessons/httpproxies/documentation/9manual.adoc
