+++ src/main/resources/lessons/httpproxies/html/HttpProxies.html
@@ -3,23 +3,23 @@
 <html xmlns:th="http://www.thymeleaf.org">
 
     <div class="lesson-page-wrapper">
-        <div class="adoc-content" th:replace="doc:lessons/http_proxies/documentation/0overview.adoc"></div>
+        <div class="adoc-content" th:replace="doc:lessons/httpproxies/documentation/0overview.adoc"></div>
     </div>
 
     <div class="lesson-page-wrapper">
-		<div class="adoc-content" th:replace="doc:lessons/http_proxies/documentation/1proxysetupsteps.adoc"></div>
+		<div class="adoc-content" th:replace="doc:lessons/httpproxies/documentation/1proxysetupsteps.adoc"></div>
 	</div>
 
     <div class="lesson-page-wrapper">
-		<div class="adoc-content" th:replace="doc:lessons/http_proxies/documentation/3browsersetup.adoc"></div>
+		<div class="adoc-content" th:replace="doc:lessons/httpproxies/documentation/3browsersetup.adoc"></div>
 	</div>
 
     <div class="lesson-page-wrapper">
-        <div class="adoc-content" th:replace="doc:lessons/http_proxies/documentation/5configurefilterandbreakpoints.adoc"></div>
+        <div class="adoc-content" th:replace="doc:lessons/httpproxies/documentation/5configurefilterandbreakpoints.adoc"></div>
     </div>
 
     <div class="lesson-page-wrapper">
-        <div class="adoc-content" th:replace="doc:lessons/http_proxies/documentation/6assignment.adoc"></div>
+        <div class="adoc-content" th:replace="doc:lessons/httpproxies/documentation/6assignment.adoc"></div>
         <div class="attack-container">
             <div class="assignment-success"><i class="fa fa-2 fa-check hidden" aria-hidden="true"></i></div>
             <form class="attack-form" accept-charset="UNKNOWN" name="intercept-request"
@@ -36,15 +36,15 @@
     </div>
 
     <div class="lesson-page-wrapper">
-        <div class="adoc-content" th:replace="doc:lessons/http_proxies/documentation/7resend.adoc"></div>
+        <div class="adoc-content" th:replace="doc:lessons/httpproxies/documentation/7resend.adoc"></div>
     </div>
     <div class="lesson-page-wrapper">
-        <div class="adoc-content" th:replace="doc:lessons/http_proxies/documentation/8httpsproxy.adoc"></div>
+        <div class="adoc-content" th:replace="doc:lessons/httpproxies/documentation/8httpsproxy.adoc"></div>
     </div>
     <div class="lesson-page-wrapper">
-        <div class="adoc-content" th:replace="doc:lessons/http_proxies/documentation/9manual.adoc"></div>
+        <div class="adoc-content" th:replace="doc:lessons/httpproxies/documentation/9manual.adoc"></div>
     </div>
     <div class="lesson-page-wrapper">
-        <div class="adoc-content" th:replace="doc:lessons/http_proxies/documentation/10burp.adoc"></div>
+        <div class="adoc-content" th:replace="doc:lessons/httpproxies/documentation/10burp.adoc"></div>
     </div>
 </html>
