+++ src/main/resources/lessons/httpproxies/i18n/WebGoatLabels.properties
