+++ src/main/resources/lessons/httpproxies/images/firefox-proxy-config.png
