+++ src/main/resources/lessons/httpproxies/images/firefoxsettingscerts.png
