+++ src/main/resources/lessons/httpproxies/images/zap-browser-button.png
