+++ src/main/resources/lessons/httpproxies/images/zap_edit_and_response.png
